backend/requirements.txt
Django==2.0.1
psycopg2-binary
djangorestframework==3.8.2
django-rest-auth==0.9.3
django-allauth==0.36.0
django-cors-headers==2.2.0
django-rest-swagger==2.2.0
uwsgi==2.0.17
django-notifs==2.6.0